Refactor routes to include a not found route
