+++ src/routes/index.ts
@@ -3,6 +3,7 @@ import { jwt_middleware } from "../services/jwt.services";
 import { rbacMiddleware } from "../services/rbac.services";
 import { authRouter } from "./v1/auth.routes";
 import homeRouter from "./v1/home";
+import { notFoundRoute } from "./v1/notFound.routes";
 import { userPublicRouter, userReadRoutes } from "./v1/users.routes";
 
 const privateRoutes = Router();
@@ -25,27 +26,5 @@ privateRoutes.use(readRoutes); // Todo! : add createRoutes, deleteRoutes  here..
 //PUBLIC ROUTES
 const publicRoutes = Router();
 publicRoutes.use(homeRouter, userPublicRouter, authRouter);
-
-//NOT FOUND ROUTE
-const notFoundRoute = Router();
-notFoundRoute.use((req, res) => {
-  const report = {
-    method: req.method,
-    url: req.url,
-    body: req.body,
-    query: req.query,
-    params: req.params,
-    headers: req.headers,
-    ip: req.ip,
-  };
-  //Todo: send the report to the admin (report)=>{send(report)}
-  res.status(404).json({
-    error: "Not Found",
-    message:
-      "The requested route does not exist, a report has been sent to the admin",
-    status: 404,
-    report,
-  });
-});
 
 export default [publicRoutes, privateRoutes, notFoundRoute];
